F2025-504-Eskala-Operations-main/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.32
PyMySQL==1.1.1
mysql-connector-python==8.0.33
python-dotenv==1.0.1
bcrypt>=4.2.0
flask-mysqldb==2.0.0
